added trait for query; phpcs on tests
