+++ src/DataMapper/Query/Delete.php
@@ -18,9 +18,7 @@ declare(strict_types=1);
 
 namespace Phalcon\DataMapper\Query;
 
-use PDOStatement;
-use Phalcon\DataMapper\Pdo\Connection;
-use Phalcon\DataMapper\Pdo\Exception\Exception;
+use Phalcon\DataMapper\Query\Traits\QueryTrait;
 use Phalcon\DataMapper\Statement\Delete as DeleteStatement;
 
 /**
@@ -28,44 +26,5 @@ use Phalcon\DataMapper\Statement\Delete as DeleteStatement;
  */
 class Delete extends DeleteStatement
 {
-    /**
-     * Create a new instance of this object
-     *
-     * @param mixed ...$arguments
-     *
-     * @return static
-     */
-    public static function new(mixed ...$arguments): static
-    {
-        $connection = Connection::new(...$arguments);
-
-        return new static($connection->getDriverName(), $connection);
-    }
-
-    /**
-     * Constructor.
-     *
-     * @param string     $driver
-     * @param Connection $connection
-     */
-    public function __construct(
-        string $driver,
-        protected Connection $connection
-    ) {
-        parent::__construct($driver);
-    }
-
-    /**
-     * Performs a statement in the connection
-     *
-     * @return PDOStatement
-     * @throws Exception
-     */
-    public function perform()
-    {
-        return $this->connection->perform(
-            $this->getStatement(),
-            $this->getBindValues()
-        );
-    }
+    use QueryTrait;
 }
