+++ src/DataMapper/Query/Insert.php
@@ -18,9 +18,7 @@ declare(strict_types=1);
 
 namespace Phalcon\DataMapper\Query;
 
-use PDOStatement;
-use Phalcon\DataMapper\Pdo\Connection;
-use Phalcon\DataMapper\Pdo\Exception\Exception;
+use Phalcon\DataMapper\Query\Traits\QueryTrait;
 use Phalcon\DataMapper\Statement\Insert as InsertStatement;
 
 /**
@@ -28,32 +26,7 @@ use Phalcon\DataMapper\Statement\Insert as InsertStatement;
  */
 class Insert extends InsertStatement
 {
-    /**
-     * Create a new instance of this object
-     *
-     * @param mixed ...$arguments
-     *
-     * @return static
-     */
-    public static function new(mixed ...$arguments): static
-    {
-        $connection = Connection::new(...$arguments);
-
-        return new static($connection->getDriverName(), $connection);
-    }
-
-    /**
-     * Constructor.
-     *
-     * @param string     $driver
-     * @param Connection $connection
-     */
-    public function __construct(
-        string $driver,
-        protected Connection $connection
-    ) {
-        parent::__construct($driver);
-    }
+    use QueryTrait;
 
     /**
      * Returns the id of the last inserted record
@@ -65,19 +38,5 @@ class Insert extends InsertStatement
     public function getLastInsertId(string $name = null): string
     {
         return $this->connection->lastInsertId($name);
-    }
-
-    /**
-     * Performs a statement in the connection
-     *
-     * @return PDOStatement
-     * @throws Exception
-     */
-    public function perform()
-    {
-        return $this->connection->perform(
-            $this->getStatement(),
-            $this->getBindValues()
-        );
     }
 }
