+++ src/DataMapper/Statement/Insert.php
@@ -188,8 +188,7 @@ class Insert extends AbstractStatement
          * understand at first glance but it is the fastest implementation.
          *
          * We are traversing the array and quoting every column.        $driver = env('driver');
-        $insert = Insert::new($driver);
-
+         * $insert = Insert::new($driver);
          */
         $columns = array_map(
             fn($column) => $this->quoteIdentifier($column),
