+++ tests/database/DataMapper/Statement/Insert/GetBindValuesTest.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Phalcon\Tests\Database\DataMapper\Statement\Insert;
 
 use PDO;
-use Phalcon\DataMapper\Statement\Delete;
 use Phalcon\DataMapper\Statement\Insert;
 use Phalcon\Tests\AbstractDatabaseTestCase;
